requirements.txt
fastapi>=0.104.0
pydantic>=2.0.0
uvicorn>=0.24.0
pyswisseph>=2.10.0
geopy>=2.4.0
timezonefinder>=6.2.0
pytz>=2023.3
httpx>=0.25.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
